+++ test/utils/const.js
@@ -1,0 +1,9 @@
+module.exports.ROOT_URL = "http://localhost:3000";
+
+module.exports.TIMEOUT = 30 * 1000;
+
+module.exports.TEST_USER = {
+  username: "testuser1",
+  email: "testuser1@example.com",
+  password: "testPassword1",
+};
